python-cron/requirements.txt
# Core dependencies
beautifulsoup4==4.12.3
requests==2.31.0
python-dotenv==1.0.1
resend==2.4.0

# Optional: for better HTML parsing
lxml==5.1.0

# Development dependencies (optional)
# pytest==8.0.0
# black==24.1.1
# flake8==7.0.0
